test2.1

+++ src/main/java/core/basesyntax/operation/PurchaseOperation.java
@@ -7,7 +7,7 @@ public class PurchaseOperation implements OperationHandler {
     @Override
     public void apply(FruitTransaction transaction) {
         String fruit = transaction.getFruit();
-        int currentQuantity = Storage.fruitStorage.getOrDefault(fruit, 0);
+        int currentQuantity = (int) Storage.fruitStorage.getOrDefault(fruit, 0);
         if (currentQuantity < transaction.getQuantity()) {
             throw new RuntimeException("Not enough " + fruit + " in stock!");
         }
